Separate app and server for easy testing
